Fixed Moff Jerjerrod not being valid on red coordinate
Never done CoffeeScript before, but it seems like the existing logic did
a boolean short-circuit, causing Moff Jerjerrod to ONLY work for a white
coordinate, and not a red one. I'm not sure if the way I fixed this is
the CoffeeScript way of doing it, but it seems to work.