Fixed another async bug
I've found one bug related to restrictions: If you Equip a Lambda-Shuttle with a targeting computer and Grand Moff Tarkin and then remove the targeting computer, an error is thrown. The reason is actually written in the comment of line 5329: validation is called twice (which is now done in parallel), so this will remove him twice at the same time.